requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
google-generativeai==0.3.2
Pillow==10.1.0
python-dotenv==1.0.0
opencv-python==4.8.1.78
numpy==1.24.3
requests==2.31.0
gunicorn==21.2.0
